+++ examples/openapi-stackexchange/CHANGELOG.md
@@ -1,5 +1,19 @@
 # openapi-stackexchange
 
+## 0.6.245
+
+### Patch Changes
+
+- Updated dependencies [f9969f116]
+  - @graphql-mesh/openapi@0.26.4
+
+## 0.6.244
+
+### Patch Changes
+
+- @graphql-mesh/cli@0.73.2
+- @graphql-mesh/openapi@0.26.3
+
 ## 0.6.243
 
 ### Patch Changes