+++ examples/openapi-subscriptions/CHANGELOG.md
@@ -1,5 +1,19 @@
 # openapi-subscriptions
 
+## 0.0.234
+
+### Patch Changes
+
+- Updated dependencies [f9969f116]
+  - @graphql-mesh/openapi@0.26.4
+
+## 0.0.233
+
+### Patch Changes
+
+- @graphql-mesh/cli@0.73.2
+- @graphql-mesh/openapi@0.26.3
+
 ## 0.0.232
 
 ### Patch Changes