+++ examples/soap-demo/CHANGELOG.md
@@ -1,5 +1,12 @@
 # soap-demo
 
+## 0.0.107
+
+### Patch Changes
+
+- @graphql-mesh/cli@0.73.2
+- @graphql-mesh/soap@0.10.63
+
 ## 0.0.106
 
 ### Patch Changes