+++ packages/cache/redis/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @graphql-mesh/cache-redis
 
+## 0.9.13
+
+### Patch Changes
+
+- Updated dependencies [d4754ad08]
+- Updated dependencies [2df026e90]
+  - @graphql-mesh/types@0.75.0
+  - @graphql-mesh/cache-localforage@0.6.13
+  - @graphql-mesh/utils@0.35.7
+
 ## 0.9.12
 
 ### Patch Changes