+++ packages/handlers/odata/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @graphql-mesh/odata
 
+## 0.17.29
+
+### Patch Changes
+
+- Updated dependencies [d4754ad08]
+- Updated dependencies [2df026e90]
+  - @graphql-mesh/types@0.75.0
+  - @graphql-mesh/store@0.8.16
+  - @graphql-mesh/utils@0.35.7
+
 ## 0.17.28
 
 ### Patch Changes