+++ packages/handlers/openapi/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @graphql-mesh/openapi
 
+## 0.26.4
+
+### Patch Changes
+
+- f9969f116: Fix serialization of array values in query parameters
+
+## 0.26.3
+
+### Patch Changes
+
+- Updated dependencies [d4754ad08]
+- Updated dependencies [2df026e90]
+  - @graphql-mesh/types@0.75.0
+  - @graphql-mesh/store@0.8.16
+  - @graphql-mesh/utils@0.35.7
+
 ## 0.26.2
 
 ### Patch Changes