+++ packages/transforms/type-merging/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @graphql-mesh/transform-type-merging
 
+## 0.4.7
+
+### Patch Changes
+
+- Updated dependencies [d4754ad08]
+- Updated dependencies [2df026e90]
+  - @graphql-mesh/types@0.75.0
+  - @graphql-mesh/utils@0.35.7
+
 ## 0.4.6
 
 ### Patch Changes