+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @graphql-mesh/utils
 
+## 0.35.7
+
+### Patch Changes
+
+- Updated dependencies [d4754ad08]
+- Updated dependencies [2df026e90]
+  - @graphql-mesh/types@0.75.0
+
 ## 0.35.6
 
 ### Patch Changes